refactor: migrate to Junit4
migrate MultiHostTestSuite

https://github.com/pgjdbc/pgjdbc/issues/205
