+++ pgjdbc/src/test/java/org/postgresql/test/hostchooser/MultiHostTestSuite.java
@@ -5,81 +5,14 @@
 
 package org.postgresql.test.hostchooser;
 
-import org.postgresql.test.TestUtil;
-import org.postgresql.util.PSQLException;
-
-import junit.framework.JUnit4TestAdapter;
-import junit.framework.TestCase;
-import junit.framework.TestSuite;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.util.Properties;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
 
 /*
  * Executes multi host tests (aka master/slave connectivity selection).
  */
-public class MultiHostTestSuite extends TestSuite {
-
-  public static Connection openSlaveDB() throws Exception {
-    TestUtil.initDriver();
-
-    Properties props = new Properties();
-
-    props.setProperty("user", TestUtil.getUser());
-    props.setProperty("password", TestUtil.getPassword());
-
-    return DriverManager.getConnection(TestUtil.getURL(getSlaveServer(), getSlavePort()), props);
-  }
-
-  /*
-   * Returns the Test server
-   */
-  public static String getSlaveServer() {
-    return System.getProperty("slaveServer", TestUtil.getServer());
-  }
-
-  /*
-   * Returns the Test port
-   */
-  public static int getSlavePort() {
-    return Integer
-        .parseInt(System.getProperty("slavePort", String.valueOf(TestUtil.getPort() + 1)));
-  }
-
-  /*
-   * The main entry point for JUnit
-   */
-  public static TestSuite suite() throws Exception {
-    TestSuite suite = new TestSuite();
-
-    if (isReplicationInstanceAvailable()) {
-      suite.addTest(new JUnit4TestAdapter(MultiHostsConnectionTest.class));
-    } else {
-      // suite must have at lest one test case
-      suite.addTestSuite(DummyTest.class);
-    }
-
-    return suite;
-  }
-
-  private static boolean isReplicationInstanceAvailable() throws Exception {
-    try {
-      Connection connection = openSlaveDB();
-      TestUtil.closeDB(connection);
-      return true;
-    } catch (PSQLException ex) {
-      return false;
-    }
-  }
-
-  public static class DummyTest extends TestCase {
-    public DummyTest(String name) {
-      super(name);
-    }
-
-    public void testDummy() {
-    }
-  }
+@RunWith(Suite.class)
+@Suite.SuiteClasses(MultiHostsConnectionTest.class)
+public class MultiHostTestSuite {
 }
